add Data_Table extraction from exd files
refactor decompression of General file

+++ src/string.h
@@ -2,7 +2,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-#define STR(x) (x).length, (x).data
+#define STR(x) (int)(x).length, (x).data
 
 typedef struct ROString {
     int64_t length;
